# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_eenemeene_kitamanager-go_internal_models.Employee:
    properties:
      birthdate:
        example: "1990-05-15"
        type: string
      contracts:
        items:
          $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.EmployeeContract'
        type: array
      created_at:
        type: string
      first_name:
        example: Max
        type: string
      gender:
        example: male
        type: string
      id:
        example: 1
        type: integer
      last_name:
        example: Mustermann
        type: string
      organization:
        $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.Organization'
      organization_id:
        example: 1
        type: integer
      section:
        $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.Section'
      section_id:
        example: 1
        type: integer
      updated_at:
        type: string
    type: object
...
  github_com_eenemeene_kitamanager-go_internal_models.EmployeeCreateRequest:
    properties:
      birthdate:
        example: "1990-05-15"
        type: string
      first_name:
        example: Max
        maxLength: 255
        type: string
      gender:
        example: male
        type: string
      last_name:
        example: Mustermann
        maxLength: 255
        type: string
      section_id:
        example: 1
        type: integer
    required:
    - birthdate
    - first_name
    - gender
    - last_name
    type: object
  github_com_eenemeene_kitamanager-go_internal_models.EmployeeUpdateRequest:
    properties:
      birthdate:
        example: "1990-05-15"
        type: string
      first_name:
        example: Max
        maxLength: 255
        type: string
      gender:
        example: male
        type: string
      last_name:
        example: Mustermann
        maxLength: 255
        type: string
      section_id:
        example: 1
        type: integer
    type: object
  github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse:
    properties:
      code:
        example: not_found
        type: string
      message:
        example: resource not found
        type: string
    type: object
...
  ? github_com_eenemeene_kitamanager-go_internal_models.PaginatedResponse-github_com_eenemeene_kitamanager-go_internal_models_Employee
  : properties:
      _links:
        $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.PaginationLinks'
      data:
        items:
          $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.Employee'
        type: array
      limit:
        example: 20
        type: integer
      page:
        example: 1
        type: integer
      total:
        example: 100
        type: integer
      total_pages:
        example: 5
        type: integer
    type: object
  ? github_com_eenemeene_kitamanager-go_internal_models.PaginatedResponse-github_com_eenemeene_kitamanager-go_internal_models_EmployeeContractResponse
  : properties:
      _links:
        $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.PaginationLinks'
      data:
        items:
          $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.EmployeeContractResponse'
        type: array
      limit:
        example: 20
        type: integer
      page:
        example: 1
        type: integer
      total:
        example: 100
        type: integer
      total_pages:
        example: 5
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@kitamanager.example.com
    name: API Support
  description: REST API for managing Users, Groups, and Organizations with RBAC support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: KitaManager API
  version: "1.0"
paths:
...
  /api/v1/organizations/{orgId}/employees:
    get:
      consumes:
      - application/json
      description: Get a paginated list of all employees in the specified organization
      parameters:
      - description: Organization ID
        in: path
        name: orgId
        required: true
        type: integer
      - description: Filter by section ID
        in: query
        name: section_id
        type: integer
      - description: Filter by active contract date (YYYY-MM-DD, defaults to today)
        in: query
        name: active_on
        type: string
      - description: Search by first or last name (case-insensitive)
        in: query
        name: search
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        maximum: 100
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.PaginatedResponse-github_com_eenemeene_kitamanager-go_internal_models_Employee'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all employees in an organization
      tags:
      - employees
    post:
      consumes:
      - application/json
      description: Create a new employee in the specified organization
      parameters:
      - description: Organization ID
        in: path
        name: orgId
        required: true
        type: integer
      - description: Employee data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.EmployeeCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.Employee'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new employee
      tags:
      - employees
  /api/v1/organizations/{orgId}/employees/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an employee by ID
      parameters:
      - description: Organization ID
        in: path
        name: orgId
        required: true
        type: integer
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete an employee
      tags:
      - employees
    get:
      consumes:
      - application/json
      description: Get a single employee by their ID
      parameters:
      - description: Organization ID
        in: path
        name: orgId
        required: true
        type: integer
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.Employee'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get employee by ID
      tags:
      - employees
    put:
      consumes:
      - application/json
      description: Update an existing employee by ID
      parameters:
      - description: Organization ID
        in: path
        name: orgId
        required: true
        type: integer
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      - description: Employee data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.EmployeeUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.Employee'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update an employee
      tags:
      - employees
  /api/v1/organizations/{orgId}/employees/{id}/contracts:
    get:
      consumes:
      - application/json
      description: Get paginated contracts for an employee
      parameters:
      - description: Organization ID
        in: path
        name: orgId
        required: true
        type: integer
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        maximum: 100
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.PaginatedResponse-github_com_eenemeene_kitamanager-go_internal_models_EmployeeContractResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List employee contracts
      tags:
      - employees
    post:
      consumes:
      - application/json
      description: |-
        Create a new contract for an employee.

        **Contract Date Rules:**
        - Both `from` and `to` dates are inclusive (the contract is active on both dates)
        - Same-day contracts are allowed (`from` == `to`)
        - Contracts must not overlap with existing contracts
        - "Touching" contracts (where contract A ends on the same day contract B starts) are considered overlapping
        - To transition between contracts, the new contract must start the day AFTER the previous one ends

        **Example:** If contract A ends on 2025-01-31, contract B must start on 2025-02-01 or later.
      parameters:
      - description: Organization ID
        in: path
        name: orgId
        required: true
        type: integer
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      - description: Contract data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.EmployeeContractCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.EmployeeContractResponse'
        "400":
          description: Invalid request (e.g., from date after to date)
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "404":
          description: Employee not found
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "409":
          description: Contract overlaps with existing contract
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create employee contract
      tags:
      - employees
  /api/v1/organizations/{orgId}/employees/{id}/contracts/{contractId}:
    delete:
      consumes:
      - application/json
      description: Delete a contract by ID
      parameters:
      - description: Organization ID
        in: path
        name: orgId
        required: true
        type: integer
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      - description: Contract ID
        in: path
        name: contractId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete employee contract
      tags:
      - employees
    get:
      consumes:
      - application/json
      description: Get a single contract by ID with properties
      parameters:
      - description: Organization ID
        in: path
        name: orgId
        required: true
        type: integer
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      - description: Contract ID
        in: path
        name: contractId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.EmployeeContractResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get employee contract by ID
      tags:
      - employees
    put:
      consumes:
      - application/json
      description: |-
        Update an existing contract by ID. The same date rules apply as for creation:
        both dates are inclusive, same-day contracts allowed, no overlapping contracts.
      parameters:
      - description: Organization ID
        in: path
        name: orgId
        required: true
        type: integer
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      - description: Contract ID
        in: path
        name: contractId
        required: true
        type: integer
      - description: Contract data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.EmployeeContractUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.EmployeeContractResponse'
        "400":
          description: Invalid request (e.g., from date after to date)
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "404":
          description: Contract not found
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "409":
          description: Updated dates would overlap with another contract
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update employee contract
      tags:
      - employees
  /api/v1/organizations/{orgId}/employees/{id}/contracts/current:
    get:
      consumes:
      - application/json
      description: Get the currently active contract for an employee
      parameters:
      - description: Organization ID
        in: path
        name: orgId
        required: true
        type: integer
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.EmployeeContractResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current employee contract
      tags:
      - employees
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"